big server class
