--- src/generators/members/index.ts
@@ -1,5 +1,0 @@
-export * from "./server-common-route-handler-method.js";
-export * from "./server-constructor.js";
-export * from "./server-per-route-handler-methods.js";
-export * from "./server-properties.js";
-export * from "./server-register-methods.js";
