--- src/generators/members/server-common-route-handler-method.ts
@@ -1,22 +1,0 @@
-import * as models from "../../models/index.js";
-import { itt } from "../../utils/iterable-text-template.js";
-import { generateCommonRouteHandlerMethodBody } from "../bodies/index.js";
-
-export function* generateServerSuperRouteHandlerMethodCode(
-  apiModel: models.Api,
-) {
-  yield* generateMethod(apiModel);
-}
-
-/**
- * generate handler for incoming requests
- */
-function* generateMethod(apiModel: models.Api) {
-  yield itt`
-    public routeHandler(
-      incomingRequest: lib.ServerIncomingRequest,
-    ): lib.ServerOutgoingResponse {
-      ${generateCommonRouteHandlerMethodBody(apiModel)}
-    }
-  `;
-}
