+++ src/generators/types/server.ts
@@ -1,12 +1,11 @@
 import * as models from "../../models/index.js";
+import { toCamel, toPascal } from "../../utils/index.js";
 import { itt } from "../../utils/iterable-text-template.js";
 import {
-  generateServerConstructorCode,
-  generateServerRouteHandleMethodsCode as generateServerPerRouteHandleMethodsCode,
-  generateServerPropertiesCode,
-  generateServerRegisterMethodsCode,
-  generateServerSuperRouteHandlerMethodCode,
-} from "../members/index.js";
+  generateCommonRouteHandlerMethodBody,
+  generateRouteHandlerMethodBody,
+} from "../bodies/index.js";
+import { generateServerConstructorBody } from "../bodies/server-constructor.js";
 
 /**
  * Generated the server class. This is the server that is generated from the
@@ -40,9 +39,113 @@ export class Server<A extends ServerAuthentication = ServerAuthentication>
 }
 
 function* generateServerBody(apiModel: models.Api) {
-  yield* generateServerPropertiesCode(apiModel);
-  yield* generateServerConstructorCode(apiModel);
-  yield* generateServerRegisterMethodsCode(apiModel);
-  yield* generateServerSuperRouteHandlerMethodCode(apiModel);
-  yield* generateServerPerRouteHandleMethodsCode(apiModel);
+  yield itt`
+    private router = new Router({
+      parameterValueDecoder: value => value,
+      parameterValueEncoder: value => value,
+    });
+  `;
+
+  for (const pathModel of apiModel.paths) {
+    for (const operationModel of pathModel.operations) {
+      const propertyName = toCamel(operationModel.name, "operation", "handler");
+      const typeName = toPascal(operationModel.name, "operation", "handler");
+
+      yield itt`
+        private ${propertyName}?: ${typeName}<A>;
+      `;
+    }
+  }
+
+  for (const authenticationModel of apiModel.authentication) {
+    const propertyName = toCamel(
+      authenticationModel.name,
+      "authentication",
+      "handler",
+    );
+    const typeName = toPascal(
+      authenticationModel.name,
+      "authentication",
+      "handler",
+    );
+
+    yield itt`
+      private ${propertyName}?: ${typeName}<A>;
+    `;
+  }
+
+  yield itt`
+    public constructor() {
+      ${generateServerConstructorBody(apiModel)}
+    }
+  `;
+
+  for (const pathModel of apiModel.paths) {
+    for (const operationModel of pathModel.operations) {
+      const methodName = toCamel("register", operationModel.name, "operation");
+      const handlerTypeName = toPascal(
+        operationModel.name,
+        "operation",
+        "handler",
+      );
+      const handlerName = toCamel(operationModel.name, "operation", "handler");
+
+      // TODO add JsDoc
+
+      yield itt`
+        public ${methodName}(operationHandler: ${handlerTypeName}<A>) {
+          this.${handlerName} = operationHandler;
+        }
+      `;
+    }
+  }
+
+  for (const authenticationModel of apiModel.authentication) {
+    const methodName = toCamel(
+      "register",
+      authenticationModel.name,
+      "authentication",
+    );
+    const handlerTypeName = toPascal(
+      authenticationModel.name,
+      "authentication",
+      "handler",
+    );
+    const handlerName = toCamel(
+      authenticationModel.name,
+      "authentication",
+      "handler",
+    );
+
+    // TODO add JsDoc
+
+    yield itt`
+      public ${methodName}(authenticationHandler: ${handlerTypeName}<A>) {
+        this.${handlerName} = authenticationHandler;
+      }
+    `;
+  }
+
+  yield itt`
+    public routeHandler(
+      incomingRequest: lib.ServerIncomingRequest,
+    ): lib.ServerOutgoingResponse {
+      ${generateCommonRouteHandlerMethodBody(apiModel)}
+    }
+  `;
+
+  for (const pathModel of apiModel.paths) {
+    for (const operationModel of pathModel.operations) {
+      const routeHandlerName = toCamel(operationModel.name, "route", "handler");
+
+      yield itt`
+        private ${routeHandlerName}(
+          routeParameters: Record<string, string>,
+          serverIncomingRequest: lib.ServerIncomingRequest,
+        ): lib.ServerOutgoingResponse {
+          ${generateRouteHandlerMethodBody(apiModel, operationModel)}
+        }
+      `;
+    }
+  }
 }
